requirements.txt
absl-py==1.3.0
accelerate==0.15.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
av==9.2.0
basicsr==1.4.2
bcrypt==4.0.1
black==22.12.0
blinker==1.5
boltons==21.0.0
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
clean-fid==0.1.29
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
decorator==5.1.1
diffusers==0.12.1
einops==0.4.1
entrypoints==0.4
facexlib==0.2.5
fairscale==0.4.9
fastapi==0.88.0
ffmpy==0.3.0
filelock==3.8.2
filterpy==1.4.5
fire==0.4.0
flatbuffers==22.12.6
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.2
gdown==4.6.0
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
gradio==3.12.0
grpcio==1.51.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.1
huggingface-hub==0.11.1
humanfriendly==10.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
inflection==0.5.1
invisible-watermark==0.1.5
Jinja2==3.1.2
joblib==1.2.0
jsonmerge==1.8.0
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.7
lark==1.1.2
librosa==0.9.2
linkify-it-py==1.0.3
llvmlite==0.39.1
lmdb==1.4.0
lpips==0.1.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.3
mypy-extensions==0.4.3
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.12.0
onnxruntime==1.13.1
open-clip-torch @ git+https://github.com/mlfoundations/open_clip.git@bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b
opencv-python==4.6.0.66
orjson==3.8.3
packaging==22.0
pandas==1.5.2
paramiko==2.12.0
pathspec==0.10.3
piexif==1.1.3
Pillow==9.3.0
platformdirs==2.6.0
pooch==1.6.0
protobuf==3.20.3
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pydeck==0.8.0
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.7.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.2.5.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
resize-right==0.0.2
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
safetensors==0.2.5
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.11.0
starlette==0.22.0
streamlit==1.18.1
streamlit-drawable-canvas==0.9.2
sympy==1.11.1
tb-nightly==2.12.0a20221208
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.7
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchdiffeq==0.2.3
torchmetrics==0.11.0
torchsde==0.2.5
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
trampoline==0.1.2
transformers==4.25.1
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
xformers==0.0.16
yapf==0.32.0
yarl==1.8.2
zipp==3.11.0
